added coverage for issue_718 test case 1
